Add support for the .js.flow file extension
The `flow` command will search for files with this extension, and report errors in them, and it seems to cause strange bugs when they are ignored by the extension.

The better solution would be to use `document.languageId` and let vscode handle the extension mapping, but this is the simplest change that fixes my issue.

This fixes #139 for me.